lightweighted main
