+++ lib/widgets/expense_item.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class ExpenseItem extends StatelessWidget {
+  final IconData icon;
+  final String label;
+  final String amount;
+
+  const ExpenseItem({
+    super.key,
+    required this.icon,
+    required this.label,
+    required this.amount,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      leading: CircleAvatar(
+        backgroundColor: Colors.green[100],
+        child: Icon(icon, color: Colors.green[800]),
+      ),
+      title: Text(label),
+      trailing: Text(
+        amount,
+        style: const TextStyle(fontWeight: FontWeight.bold),
+      ),
+    );
+  }
+}
