+++ lib/widgets/tab_item_data.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class TabItemData {
+  final IconData? icon;
+  final String label;
+  final String? assetPath;
+  const TabItemData({this.icon, required this.label, this.assetPath});
+}
